Extract API logic to separate class
