+++ conf/application.conf
@@ -1,4 +1,5 @@
 twitter_api {
     auth_bearer_token="<PLACEHOLDER>"
     tweet_keywords_query=["covid", "covid19", "coronavirus"]
+    tweet_fields=["id", "text", "created_at", "lang"]
 }
