+++ src/main/scala/Main.scala
@@ -1,5 +1,6 @@
 import collection.JavaConverters._
 import config.ConfigManager
+import api.TwitterAPIClient
 import requests.get
 import ujson.read
 
@@ -7,22 +8,14 @@ import ujson.read
 def main(args: String*): Unit = {
   // TODO: refactor, modularise, handle exceptions
   val conf = ConfigManager.config
-
-  val tweetKeywords = conf.getStringList("twitter_api.tweet_keywords_query").asScala.toList
-  val searchQuery = (tweetKeywords ++ tweetKeywords.map(t => s"#$t")).mkString(" OR ")
-
   val bearerToken = conf.getString("twitter_api.auth_bearer_token")
-  val response = requests.get(
-    url = "https://api.twitter.com/2/tweets/search/recent",
-    params = Map(
-      "query" -> searchQuery,
-      "tweet.fields" -> "id,text,created_at,lang",
-    ),
-    headers = Map("Authorization" -> s"Bearer $bearerToken")
-  )
-  val response_data = ujson.read(response.text())("data").arr
+  val tweetKeywords = conf.getStringList("twitter_api.tweet_keywords_query").asScala.toList
+  val tweetFields = conf.getStringList("twitter_api.tweet_fields").asScala.toList
 
-  response_data.foreach(
+  val twitterApiClient = new TwitterAPIClient(bearerToken)
+  val tweets = twitterApiClient.getTweets(tweetKeywords, tweetFields)
+
+  tweets.foreach(
     tweet => println(
       s"Tweet id: ${tweet("id")}\n" +
       s"Tweet text: ${tweet("text")}\n" +
